+++ esp-wifi/src/wifi/os_adapter.rs
@@ -1497,7 +1497,7 @@ pub unsafe extern "C" fn log_writev(
 ///
 /// *************************************************************************
 pub unsafe extern "C" fn log_timestamp() -> u32 {
-    crate::current_millis() as u32
+    esp_hal::time::now().duration_since_epoch().to_millis() as u32
 }
 
 /// **************************************************************************